# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Admin App Documentation",
        "description": "Laravel Admin App OpenApi description",
        "contact": {
            "email": "abiolafakorede@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Laravel Admin API Server"
        }
    ],
    "paths": {
        "/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns collection of users",
                "operationId": "getUsersList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Pagination Page",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a User",
                "description": "Creates a User",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get user by id",
                "description": "Returns a single user",
                "operationId": "getSingleUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a User",
                "description": "Updates a User",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a User",
                "description": "Delete a User by Id",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/user": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get User Details",
                "description": "Returns details of authenticated user",
                "operationId": "getUserDetails",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/info": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update User Details",
                "description": "Updates Authenticated User Details",
                "operationId": "updateUserInfo",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateInfoRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/password": {
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update User Password",
                "description": "Updates Authenticated User Password",
                "operationId": "updateUserPassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "UpdateInfoRequest": {
                "title": "Update User Info Request",
                "description": "Update Authenticated User Info request body data",
                "properties": {
                    "first_name": {
                        "title": "first_name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "last_name",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdatePasswordRequest": {
                "title": "Update User Password Request",
                "description": "Update Authenticated User Password request body data",
                "properties": {
                    "password": {
                        "title": "password",
                        "type": "string"
                    },
                    "password_confirm": {
                        "title": "password_confirm",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UserCreateRequest": {
                "title": "Store User Request",
                "description": "Store user request body data",
                "properties": {
                    "first_name": {
                        "title": "first_name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "last_name",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "type": "string"
                    },
                    "role_id": {
                        "title": "role_id",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "UserUpdateRequest": {
                "title": "Update User Request",
                "description": "Update user request body data",
                "properties": {
                    "first_name": {
                        "title": "first_name",
                        "type": "string"
                    },
                    "last_name": {
                        "title": "last_name",
                        "type": "string"
                    },
                    "email": {
                        "title": "email",
                        "type": "string"
                    },
                    "role_id": {
                        "title": "role_id",
                        "type": "integer"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}